trend_agent/requirements.txt
cloudpickle==3.1.1
db-dtypes==1.4.3
fastapi==0.116.1
google-adk==1.12.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth-httplib2==0.2.0
google-auth==2.40.3
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.36.0
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-genai==1.31.0
numpy==2.3.2
pandas==2.3.2
tabulate==0.9.0
uvicorn==0.35.0